Ensure feature flagging works for the dummy app (including tests).
The dummy app of this addon is transpiled by the "project", not the
ember-data addon this means that the special
`@ember-data/canary-features` processing that we do for each of our
project's `addon/**` folders does not apply to the dummy app (which
includes tests!).

This ensures that the same `@ember-data/canary-features` processing that
the various ember-data addons do is done in the dummy app.